log duplicated KSS references instead of replacing them
The original logic worked great, but in the internal repo it seems
that the external `typography.fonts` was overwriting the internal one
(external came later). So rather than replacing, we'll simply log a
message if we encounter a duplicate. It'll be up to the dev to ensure
that things are being rendered correctly.