sass "bundle" option to copy assets (#195)
* remove postcss-import and postcss-copy-assets

as cool as they are, i would prefer instead to _not_ .npmignore the original `resources/` directory. this is particularly important for files such as icons.json that aren't referenced in the source code but should be included in the installed package.

* remove resources/ from .npmignore files

* actually bring back those postcss plugins

but hide them behind an option in the sass task. this is because some projects actually want this behavior.

* docs + landing projects bundle sass assets

also alphabetize options and add project "interface" documentation.