Adjust Truncated Popover Behavior
Previously, the `TruncatedFormat` cell would display
a popover target with the "..." icon when its text
content was longer than the `truncationLength` prop.

Cells that weren't truncated this way used a CSS
"text-overflow: ellipsis" rule, which lead to
the odd sight of a mixture of "..." in the cells.

This change alters the default behavior to always
show the large "..." icon even if the cell's text
content isn't truncated. Therefore the look of the
cells will be much more consistent.

Also, the `JSONFormat` cell will turn off the
"..." popover target for nully values.