Force Karma to serve ts/tsx files as typescript (#409)
The default mime type of these files is a QuickTime video format. When
running unit tests in a Chrome browser, strict MIME type checking would
prevent the tests from running. This change allows us to serve `*.ts`
files to a Karma server and allow them to be executed in a Chrome
browser. PhantomJS doesn't seem have strict mime-type checking, so this
doesn't affect tests on CircleCI.