rewrap docs comments to 100 characters
not because we have to, because we _get to_!

I used Rewrap extension in VS Code (https://marketplace.visualstudio.com/items?itemName=stkb.rewrap) to do all the heavy lifting, with some manual wrapping as necessary.
Length isn't linted in comments but it's easy to maintain with an editor extension.