Fix collapse.tsx for TS 2.4 (#1296)
React.HTMLAttributes requires `overflowY` to be a member of a string
union: `"initial" | "inherit" | "unset" | "auto" | "hidden" | "scroll" |
"visible" | undefined`. In Collapse.render, the `overflowY` needs to be
given the explicit type `"visible" | undefined` in order to be
assignable to HTMLAttributes. Otherwise, it gets the type `string`,
which is too general.