Allow edits to be rejected for EditableCell
Currently, if I don't change the controlled value as the result of a `onConfirm`, this check prevents the original value from overriding the savedValue. This is useful in a case where input to the `EditableCell` is being validated, and if it doesn't pass validation, the value of the cell shouldn't change.