Fix a controlled EditableText bug
Fixes a bug where EditableText controlled value cannot be reset to undefined or null.

The comments in this file state that the point of mirroring the `props.value` to `state.value` is to maintain that value while the user
is editing. However, if the user is not editing, the edited line in question as is does not allow for the value of the component to be
reset to something undefined or null for emptiness.