Set aria-live on Overlay
This allows screen readers to read content in Tooltips, Popovers,
Dialogs, and Overlays when they are opened.

Not using aria-describedby because the Overlay content is typically
added to the DOM after interaction with some other element, leaving
screen readers unable to consistently see the content in time. Setting
aria-live on the Overlay works because it remains mounted after the
initial open. VoiceOver is also able to read the content on initial
open (did not test with other screen readers).