Focus input when closing DateInput popover
Since the popover is usually open in a portal which makes it one of the
last elements in the DOM, when closing the popover with tab or
shift+tab, focus is likely to go to the browser address bar or an
element at the bottom of the page, respectively.

This change adds invisible divs to the beginning and end of the popover
that, when focused, return focus to the input and close the popover. We
cannot use the existing Overlay#shouldReturnFocusOnClose prop because it
would reopen the popover.