Stop preventing default action on keydown
Previously we would replay any clicks on keyup, but this was bad because
keyup is not guaranteed to run on the same element. Better to just let
the default click occur on keydown and stop fighting the browser. Now
that we no longer click on keyup, we need to also stop preventing
default on keydown to allow the native event to flow through.