popover2: allow overriding aria-haspopup
This adds a new `popupKind` property to the `Popover2` component for
controlling the `aria-haspopup` attribute of the target element.

The current implementation hard-codes `aria-haspopup` to "true" which
means that the popup is a menu. However, this is not true in all cases.
Furthermore, the `aria-haspopup` attribute should not be set for
tooltips since they are not interactive elements. So if the interaction
kind is a hover interaction, this new property is ignored.

Reference: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-haspopup