Update first-mate to use oniguruma 7.2.3
Fixes the build on macOS when using Node 16.14+ on macOS. Previously,
the build would fail with the following error:

error: no template named 'remove_cv_t' in namespace 'std'; did you mean 'remove_cv'

It required contributors to downgrade to 14.17.x.

This bug was fixed in oniguruma@7.2.3. See
https://github.com/atom/node-oniguruma/pull/112. first-mate@7.4.3 is
the first to require oniguruma@7.2.3.